[v2] Add Integration Tests (#326)
Following bugs where found and fixed:

* handling paging correctly
* handling duration and temporal scalars correctly
* fixing union connection issues
* using correct logical operator for `SOME` and `SINGLE` relation predicates

The test-framework is extended by the following features:

* validating query results via JsonPath in asciidoc
* testing custom resolvers via asciidoc
* using neo4j-driver connection for integration tests

+++ core/src/main/kotlin/org/neo4j/graphql/scalars/BigIntScalar.kt
@@ -23,8 +23,8 @@ object BigIntScalar {
             override fun parseValue(input: Any, graphQLContext: GraphQLContext, locale: Locale): Number? {
                 return when (input) {
                     is StringValue -> input.value.toLong()
-                    is FloatValue -> input.value
-                    is IntValue -> input.value
+                    is FloatValue -> input.value.toLong()
+                    is IntValue -> input.value.toLong()
                     is String -> input.toLong()
                     is Float, is Int, is Long -> input as Number
                     else -> Assert.assertShouldNeverHappen("Only string or number is expected")
